This is the latest version of lm studio web chat
This is the latest version of Lm Studio Web Chat, a dark theme like VSC or CURSOR. It has the ability to select an IP and port, which LM STUDIO CHAT lacks. Chat history can be saved and deleted. The LLM model can be loaded remotely on the LM STUDIO server. The final version will be released soon. Code formatting will be added. The program is fully functional.